Fixed multiple issues when dealing with projects with multiple language folders.
Changed folder searching using the show folder setting instead of hard coded values
Updated the table layout to handle folders better
Fixed the table layout to not show fields for missing files in a folder.
Updated the table layout to show the default for the given folder.
Fixed the headings in a vertical position